fix: X3-X7 database hardening batch
X3: 4 composite indexes (edge_zones_ab, edges_last_seen_count, edges_b_a, fusions_ab)
X4: PRAGMA user_version schema versioning + idempotent migration
X5: days_to_date() fallback returns UTC today, not hardcoded 2026-01-01
X6: saturation_loss float (was int, crashed count=1000 to 1)
X7: feed progress order verified (save before progress write)
8 tests, all PASS.

+++ engine/core/mycelium.py
@@ -853,7 +853,7 @@ class Mycelium:
                 new_count = count / (2 ** periods)
 
                 if (self.SATURATION_BETA > 0 and new_count > self.SATURATION_THRESHOLD):
-                    saturation_loss = int(self.SATURATION_BETA * new_count * new_count)
+                    saturation_loss = self.SATURATION_BETA * new_count * new_count
                     new_count = max(1, new_count - saturation_loss)
 
                 if new_count < 0.01:
@@ -890,7 +890,7 @@ class Mycelium:
                     periods = age_days // days
                     new_count = conn["count"] / (2 ** periods)
                     if (self.SATURATION_BETA > 0 and new_count > self.SATURATION_THRESHOLD):
-                        saturation_loss = int(self.SATURATION_BETA * new_count * new_count)
+                        saturation_loss = self.SATURATION_BETA * new_count * new_count
                         new_count = max(1, new_count - saturation_loss)
                     if new_count < 0.01:
                         dead.append(key)
